+++ site/package.json
@@ -2,10 +2,10 @@
   "name": "site",
   "version": "0.0.1",
   "dependencies": {
-    "@code-hike/mini-browser": "0.0.1",
-    "@code-hike/mini-editor": "0.0.1",
-    "@code-hike/mini-terminal": "0.0.1",
-    "@code-hike/scroller": "0.0.1",
+    "@code-hike/mini-browser": "0.2.1",
+    "@code-hike/mini-editor": "0.2.1",
+    "@code-hike/mini-terminal": "0.2.1",
+    "@code-hike/scroller": "0.2.1",
     "@code-surfer/standalone": "^3.1.1",
     "@emotion/core": "10.0.7",
     "@mdx-js/loader": "^1.6.5",
